refactor mysql_charsets.inc.php
