+++ db_create.php
@@ -10,7 +10,7 @@
  */
 require_once 'libraries/common.inc.php';
 
-require_once 'libraries/mysql_charsets.lib.php';
+require_once 'libraries/mysql_charsets.inc.php';
 if (! PMA_DRIZZLE) {
     include_once 'libraries/replication.inc.php';
 }
