+++ db_routines.php
@@ -11,7 +11,7 @@
  */
 require_once 'libraries/common.inc.php';
 require_once 'libraries/Util.class.php';
-require_once 'libraries/mysql_charsets.lib.php';
+require_once 'libraries/mysql_charsets.inc.php';
 
 /**
  * Include all other files
