+++ libraries/display_create_database.lib.php
@@ -27,7 +27,7 @@ if ($is_create_db_priv) {
             <input type="hidden" name="reload" value="1" />
             <input type="text" name="new_db" value="<?php echo $db_to_create; ?>" maxlength="64" class="textfield" id="text_create_db"/>
     <?php
-    include_once './libraries/mysql_charsets.lib.php';
+    include_once './libraries/mysql_charsets.inc.php';
     echo PMA_generateCharsetDropdownBox(PMA_CSDROPDOWN_COLLATION, 'db_collation', null, null, true, 5);
 
     if (! empty($dbstats)) {
