+++ libraries/display_structure.lib.php
@@ -7,7 +7,7 @@
  */
 
 require_once 'libraries/common.inc.php';
-require_once 'libraries/mysql_charsets.lib.php';
+require_once 'libraries/mysql_charsets.inc.php';
 require_once 'libraries/structure.lib.php';
 require_once 'libraries/index.lib.php';
 require_once 'libraries/tbl_info.inc.php';
