+++ libraries/sqlparser.lib.php
@@ -44,7 +44,7 @@ require_once './libraries/sqlparser.data.php';
  * Charset information
  */
 if (!defined('TESTSUITE')) {
-    include_once './libraries/mysql_charsets.lib.php';
+    include_once './libraries/mysql_charsets.inc.php';
 }
 if (! isset($mysql_charsets)) {
     $mysql_charsets = array();
