+++ tbl_select.php
@@ -13,7 +13,7 @@
  * Gets some core libraries
  */
 require_once 'libraries/common.inc.php';
-require_once 'libraries/mysql_charsets.lib.php';
+require_once 'libraries/mysql_charsets.inc.php';
 require_once 'libraries/TableSearch.class.php';
 
 $response = PMA_Response::getInstance();
