+++ test/libraries/rte/PMA_RTN_getEditorForm_test.php
@@ -13,7 +13,7 @@ require_once './libraries/Types.class.php';
 require_once 'libraries/Theme.class.php';
 require_once 'libraries/database_interface.inc.php';
 require_once 'libraries/Tracker.class.php';
-require_once 'libraries/mysql_charsets.lib.php';
+require_once 'libraries/mysql_charsets.inc.php';
 /*
  * Include to test.
  */
